people: profile visible for owner even if the owner set it to private.

+++ src/www/people/viewprofile.php
@@ -51,7 +51,14 @@ if ($user_id && is_numeric($user_id)) {
 		/*
 			profile set private
 		*/
-		if (db_result($result, 0, 'people_view_skills') != 1) {
+		$overwritten_access = 0;
+		if (session_loggedin()) {
+			$u = user_get_object(user_getid());
+			if ($u->getID() == $user_id) {
+				$overwritten_access = 1;
+			}
+		}
+		if ((db_result($result, 0, 'people_view_skills') != 1) && !$overwritten_access) {
 			echo '<p class="warning">'._('This user has set his/her profile to private.').'</p>';
 			people_footer(array());
 			exit;
